move ID updater function to separate module
